requirements.txt
alembic==1.13.2
astroid==2.9.3
attrs==21.2.0
Automat==20.2.0
Babel==2.8.0
bcrypt==3.2.0
blinker==1.8.2
certifi==2020.6.20
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.7
cloud-init==24.3.1
colorama==0.4.4
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==3.4.8
dbus-python==1.2.18
distlib==0.3.8
distro==1.7.0
distro-info==1.1+ubuntu0.2
exceptiongroup==1.2.2
filelock==3.16.1
Flask==3.0.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
greenlet==3.1.0
httplib2==0.20.2
hyperlink==21.0.0
idna==3.10
importlib-metadata==4.6.4
incremental==21.3.0
iniconfig==2.0.0
isort==5.6.4
itsdangerous==2.2.0
jeepney==0.7.1
Jinja2==3.1.4
jsonpatch==1.32
jsonpointer==2.0
jsonschema==3.2.0
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy-object-proxy==0.0.0
logilab-common==1.8.2
Mako==1.3.5
MarkupSafe==2.1.5
mccabe==0.6.1
more-itertools==8.10.0
mypy-extensions==0.4.3
netifaces==0.11.0
oauthlib==3.2.0
packaging==24.1
pipenv==2024.0.2
pipfile==0.0.2
platformdirs==4.3.6
pluggy==1.5.0
psycopg2-binary==2.9.9
pyasn1==0.4.8
pyasn1-modules==0.2.1
pycurl==7.44.1
PyGObject==3.42.1
PyHamcrest==2.0.2
PyJWT==2.3.0
pylint==2.12.2
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyrsistent==0.18.1
pyserial==3.5
pytest==8.3.2
python-apt==2.4.0+ubuntu4
pytz==2024.2
PyYAML==5.4.1
requests==2.32.3
SecretStorage==3.3.1
service-identity==18.1.0
six==1.16.0
SQLAlchemy==2.0.29
sqlalchemy-serializer==1.4.22
systemd-python==234
toml==0.10.2
tomli==2.0.1
Twisted==22.1.0
typing_extensions==4.12.2
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==2.2.3
virtualenv==20.26.5
virtualenv-clone==0.3.0
wadllib==1.3.6
Werkzeug==3.0.4
wrapt==1.13.3
zipp==1.0.0
zope.interface==5.4.0
